add: adds all neccessary functions
With this commit, I added basic functionality for my app. You can now not only choose the password's length, but also the characters used. Also, a check is added to make sure at least one character of each character set is being used. The code still needs to be refined.